requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
python-dotenv
pydantic>=2,<3
psycopg2-binary
